requirements.txt
alembic==1.8.0
attrs==22.1.0
awsebcli==3.20.3
bcrypt==3.2.2
blessed==1.19.1
blinker==1.4
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
cryptography==37.0.4
dnspython==2.2.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.16.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.0.1
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==3.2.0
Mako==1.2.0
MarkupSafe==2.1.1
paramiko==2.11.0
pathspec==0.9.0
Pillow==9.1.1
psycopg2-binary==2.9.3
py3dns==3.2.1
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
SQLAlchemy==1.4.37
termcolor==1.1.0
texttable==1.6.4
urllib3==1.26.11
validate-email-address==1
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
